requirements.txt
Django==2.2.19
Pillow==8.3.1
djangorestframework==3.12.4
